+++ HISTORY.rst
@@ -6,6 +6,7 @@ History
 HEAD
 ----
 
+- Adding adapter and quality JSON fields to ``FlowCell`` and APIs to set them
 - Some layout / UI refinements.
 - Adding unstable API, mostly read-only except for what is needed for automated demultiplexing and QC.
 - Adding UI for generating REST API login tokens.